updated GUI

+++ app/src/main/java/com/example/mainaccount/inspire/adapters/FavoriteListAdapter.java
@@ -1,6 +1,7 @@
 package com.example.mainaccount.inspire.adapters;
 
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,10 +11,13 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.example.mainaccount.inspire.R;
+import com.example.mainaccount.inspire.model.SigninActivity;
+import com.google.firebase.auth.FirebaseAuth;
 
 import java.util.List;
 
 import static android.content.Context.MODE_PRIVATE;
+import static com.example.mainaccount.inspire.model.BaseActivity.userIntent;
 
 /**
  *  Classname: FavoriteListAdapter.java
@@ -29,12 +33,17 @@ public class FavoriteListAdapter extends BaseAdapter {
     SharedPreferences sharedPreferences;
     SharedPreferences.Editor editor;
     private String delete;
+    Intent intent;
 
     public FavoriteListAdapter(Context mContext, List<String> gemList) {
         this.mContext = mContext;
         this.gemList = gemList;
         sharedPreferences = mContext.getSharedPreferences(MyFavPREFERENCES, MODE_PRIVATE);
         editor = sharedPreferences.edit();
+    }
+
+    public void setIntent(Intent intent){
+        this.intent = intent;
     }
 
     @Override
@@ -61,8 +70,6 @@ public class FavoriteListAdapter extends BaseAdapter {
         tvFav.setText(gemList.get(position).toString());
         delete = gemList.get(position).toString();
 
-
-
         Button deleteFavBtn = (Button) v.findViewById(R.id.delete_fav_btn);
         Button deleteAllBtn = (Button) v.findViewById(R.id.delete_all_btn);
         deleteAllBtn.setVisibility(View.INVISIBLE);
@@ -74,14 +81,34 @@ public class FavoriteListAdapter extends BaseAdapter {
         deleteFavBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                // delete favorite record from SharedPreferences storage
-                editor.remove(delete);
-                editor.commit();
-                // updtate ListView of change in data
-                gemList.remove(position);
-                notifyDataSetChanged();
-                // inform user of change
-                Toast.makeText(mContext, "Favorite deleted!", Toast.LENGTH_SHORT).show();
+                if(FirebaseAuth.getInstance().getCurrentUser() != null) {
+                    // delete favorite record from SharedPreferences storage
+                    editor.remove(delete);
+                    editor.commit();
+                    // updtate ListView of change in data
+                    gemList.remove(position);
+                    notifyDataSetChanged();
+                    // inform user of change
+                    Toast.makeText(mContext, "Favorite deleted!", Toast.LENGTH_SHORT).show();
+                }else{
+                    Toast.makeText(mContext, "You must be signed in to delete a favorite!\n" +
+                            "Redirecting to sign in...", Toast.LENGTH_LONG).show();
+                    userIntent = intent;
+                    Thread thread = new Thread(){
+                        @Override
+                        public void run() {
+                            try {
+                                Thread.sleep(2500); // Launch login Activity after Toast message has run
+                                Intent i = new Intent(mContext, SigninActivity.class);
+                                viewGroup.getContext().startActivity(i);
+                            } catch (Exception e) {
+                                e.printStackTrace();
+                            }
+                        }
+                    };
+
+                    thread.start();
+                }
 
 
             }
@@ -91,14 +118,34 @@ public class FavoriteListAdapter extends BaseAdapter {
         deleteAllBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                // delete all favorite records from SharedPreferences storage
-                editor.clear();
-                editor.commit();
+                if(FirebaseAuth.getInstance().getCurrentUser() != null) {
+                    // delete all favorite records from SharedPreferences storage
+                    editor.clear();
+                    editor.commit();
 
-                // updtate ListView of change in data
-                gemList.clear();
-                notifyDataSetChanged();
-                Toast.makeText(mContext, "Favorites list deleted!", Toast.LENGTH_SHORT).show();
+                    // updtate ListView of change in data
+                    gemList.clear();
+                    notifyDataSetChanged();
+                    Toast.makeText(mContext, "Favorites list deleted!", Toast.LENGTH_SHORT).show();
+                }else{
+                    Toast.makeText(mContext, "You must be signed in to delete favorites!\n" +
+                            "Redirecting to sign in...", Toast.LENGTH_LONG).show();
+                    userIntent = intent;
+                    Thread thread = new Thread(){
+                        @Override
+                        public void run() {
+                            try {
+                                Thread.sleep(2500); // Launch login Activity after Toast message has run
+                                Intent intent = new Intent(mContext, SigninActivity.class);
+                                viewGroup.getContext().startActivity(intent);
+                            } catch (Exception e) {
+                                e.printStackTrace();
+                            }
+                        }
+                    };
+
+                    thread.start();
+                }
 
             }
         });
